Refactor before adding second class
